Add "Color Hint Background" option
Some search engines and extensions allow users to define custom colours
for their search pages. Users with custom colorization may find that the
color hints in the popup interface are inaccurate. Add an option to
customize the background of the hints so those users can match them to
their custom search page colours.